Issue #37 : On editing a post, scroll partial to top of preview.
Editing a post field will trigger a refresh of its partial.
In the refresh() method, define and call scrollPartialToTopOfPage.
Some themes have a margin-top of the #page element, like 2016.
And this scrolling would hide part of the partial.
So also calculate the offsetTop of #page, to prevent this.
If this isn't a concern, this commit could be much simpler.