Ensure that preview urls are used as permalinks for customized posts
This fixes an issue in snapshots frontend preview where auto-draft posts previewed as published would get erroneously redirected from the preview URL to instead have a URL like /-0001/11/30/ since the date is 0000-00-00 and the post_name is empty. This URL would then redirect to the homepage.

This also ensures that WordPress will be able to locate the customized post since the post_name may have been changed.